Split activity details schema
