+++ src/marine-licences/models/site-details/site-details.js
@@ -3,7 +3,7 @@ import { marineLicenceId } from '../shared-models.js'
 import { coordinatesTypeFieldSchema } from '../../../exemptions/models/site-details/coordinates-type.js'
 import { fileUploadConditionalSiteItemFields } from '../../../shared/models/site-details/file-upload.js'
 import { siteNameFieldSchema } from '../../../shared/models/site-details/site-name.js'
-import { activityItemSchema } from '../activity-details.js'
+import { activityItemSchema } from '../activity-details/activity-details.js'
 
 export const siteItemSchema = joi.object({
   coordinatesType: coordinatesTypeFieldSchema,
